refactor: moved around types
