+++ src/models/category.go
@@ -1,0 +1,9 @@
+package types
+
+type CatRequest struct {
+	Category string `json:"category"`
+}
+
+type CatList struct {
+	Categories []string `json:"categories"`
+}
