+++ src/service/downloads.go
@@ -1,7 +1,9 @@
-package mam
+package service
 
 import (
+	"errors"
 	"fmt"
+	"github.com/RA341/gouda/download_clients"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/viper"
 	"io"
@@ -11,9 +13,79 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"runtime"
+	"time"
 )
 
-// anon mouse related functions
+func ProcessDownloads(clientPointer *download_clients.DownloadClient, torrentId string, author, book, category string) {
+	downloadClient := *clientPointer
+
+	timeRunning := time.Second * 0
+	timeout := time.Minute * viper.GetDuration(fmt.Sprintf("download.timeout"))
+
+	for {
+		log.Info().Msgf("getting torrent info with id:%s", torrentId)
+		info, err := downloadClient.CheckTorrentStatus(torrentId)
+		if err != nil {
+			log.Warn().Msgf("Failed to get info with id: %s", torrentId)
+			break
+		} else {
+			if info.Status == "seeding" {
+				destPath := viper.GetString("folder.defaults")
+				catPath := fmt.Sprintf("%s/%s", destPath, category)
+				destPath = fmt.Sprintf("%s/%s/%s", catPath, author, book)
+
+				info.DownloadPath = fmt.Sprintf("%s/%s", info.DownloadPath, info.Name)
+
+				log.Info().Msgf("Torrent %s complete, linking to download location: %s ------> destination: %s", info.Name, info.DownloadPath, destPath)
+
+				err := HardLinkFolder(info.DownloadPath, destPath)
+				if err != nil {
+					log.Error().Err(err).Msgf("Failed to link info with id: %d", torrentId)
+					break
+				}
+
+				sourceUID := viper.GetInt("user.uid")
+				sourceGID := viper.GetInt("user.gid")
+				log.Info().Msgf("Changing file perm to %d:%d", sourceUID, sourceGID)
+				err = recursiveChown(catPath, sourceUID, sourceGID)
+				if err != nil {
+					log.Error().Err(err).Msgf("Failed to chown info with id: %d", torrentId)
+					break
+				}
+
+				log.Info().Msgf("Download complete: %s and hardlinked to %s", info.Name, destPath)
+				break
+			}
+
+			if timeout < timeRunning {
+				log.Error().Msgf("Maximum timeout reached abandoning check for %d:%s after %s, max timeout was set to %s, check your torrent downloadClient or increase timeout in settings", torrentId, info.Name, timeRunning.String(), timeout)
+				break
+			}
+
+			timeRunning = timeRunning + 1*time.Minute
+			log.Info().Msgf("Torrent check for %s with status:%s checking again in a minute", info.Name, info.Status)
+			time.Sleep(1 * time.Minute)
+		}
+	}
+}
+
+func recursiveChown(path string, uid, gid int) error {
+	// Walk the directory tree
+	return filepath.Walk(path, func(name string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+
+		// Change ownership of the current file/directory
+		err = os.Chown(name, uid, gid)
+		if err != nil {
+			return fmt.Errorf("failed to chown %s: %v", name, err)
+		}
+
+		return nil
+	})
+}
 
 // HardLinkFolder creates hard links of all files from source folder to target folder
 // and maintains the original UID/GID
@@ -92,6 +164,10 @@ func HardLinkFolder(sourceDir, targetDir string) error {
 
 // createHardLink creates a hard link with proper ownership and permissions
 func createHardLink(oldPath, newPath string, uid, gid int, mode fs.FileMode) error {
+	if runtime.GOOS == "windows" {
+		return errors.New("hardlinking on windows is not supported")
+	}
+
 	// Ensure the target directory exists
 	targetDir := filepath.Dir(newPath)
 	if err := os.MkdirAll(targetDir, 0755); err != nil {
